allow more control over output, use stac catalog api v1

+++ py/stac2video.py
@@ -12,6 +12,7 @@ import numpy
 import pystac_client
 import stackstac
 import xarray
+from PIL import Image
 from rasterio.enums import Resampling
 
 # 0 NO_DATA
@@ -28,15 +29,21 @@ from rasterio.enums import Resampling
 # 11 SNOW
 CLOUDY_OR_NODATA = (0, 3, 8, 9, 10)
 
-catalog = pystac_client.Client.open("https://earth-search.aws.element84.com/v0/")
+catalog = pystac_client.Client.open("https://earth-search.aws.element84.com/v1/")
 
-COLLECTION = "sentinel-s2-l2a-cogs"
+COLLECTION = "sentinel-2-l2a"
+
+NODATA = 0
 
 tms = morecantile.tms.get("WebMercatorQuad")
 
 
 def fetch_composites(
-    tile: morecantile.commons.Tile, start: str, end: str, interval: str = "14D"
+    tile: morecantile.commons.Tile,
+    start: str,
+    end: str,
+    interval: str = "14D",
+    up: int = 0,
 ) -> xarray.DataArray:
     search = catalog.search(
         collections=[COLLECTION],
@@ -46,29 +53,27 @@ def fetch_composites(
 
     items = search.get_all_items()
 
-    print(f"Found {len(items)} items")
-
     stack = stackstac.stack(
         items,
         bounds=tms.xy_bounds(tile),
         snap_bounds=False,
         epsg=tms.crs.to_epsg(),
-        resolution=tms._resolution(tms.matrix(zoom=tile.z)),
+        resolution=tms._resolution(tms.matrix(zoom=tile.z + up)),
         dtype="uint16",
         fill_value=0,
         resampling=Resampling.bilinear,
         assets=[
-            "B02",
-            "B03",
-            "B04",
-            "SCL",
+            "blue",
+            "green",
+            "red",
+            "scl",
         ],
     )
 
     data = stack.compute()
 
     # Cloud mask
-    scl = data.sel(band="SCL").astype("uint8")
+    scl = data.sel(band="scl").astype("uint8")
 
     cloud_mask = scl.isin(CLOUDY_OR_NODATA)
 
@@ -137,6 +142,19 @@ def tile_center(tile: morecantile.commons.Tile) -> Tuple[float]:
     "--interval",
     default="14D",
     help="Time interval to use for compositing imagery. Defaults to 14D. This will be passed to the `resample` function on the time dimension from xarray. See https://docs.xarray.dev/en/stable/generated/xarray.DataArray.resample.html",
+    type=str,
+)
+@click.option(
+    "--up",
+    default=0,
+    type=int,
+    help="Use higher zoom level to determine resolution of output videos. The default resolution is 256x256 pixels. Up works as multiplier. For instance, for up=1, the resolution of the video will be 512x512 pixels, and up=2 it will be 1024x1024 px",
+)
+@click.option("--images", is_flag=True, help="Output each frame as png image as well")
+@click.option(
+    "--keep-mp4",
+    is_flag=True,
+    help="Keep intermediary mp4 files",
 )
 def stac_tile(
     dst: Path,
@@ -148,7 +166,10 @@ def stac_tile(
     width: int,
     height: int,
     interval: str,
-):
+    up: int,
+    images: bool,
+    keep_mp4: bool,
+) -> None:
     if not dst.exists():
         raise ValueError(
             f"Target folder {dst} does not exist, please create it before running this script."
@@ -168,13 +189,13 @@ def stac_tile(
         "name": "Videomap",
         "base_url": "videos/",
         "center": tile_center(orig),
-        "zoom": zoom - 1,
+        "zoom": zoom,
         "frames": {},
         "features": [],
     }
 
     for tile in tiles:
-        composites = fetch_composites(tile, start, end, interval)
+        composites = fetch_composites(tile, start, end, interval, up)
 
         filepath = dst / f"videomap-{tile.z}-{tile.x}-{tile.y}.mp4"
         filepathwebm = dst / f"videomap-{tile.z}-{tile.x}-{tile.y}.webm"
@@ -186,18 +207,33 @@ def stac_tile(
             fps=4,
             frameSize=composites.shape[-2:],
         )
+        print(f"Creating videos with frame size of {composites.shape[-2:]}")
         for img in composites:
+            # Prepare video data
             bgr = img.data[:3].transpose(1, 2, 0)
-            bgr = 255 * bgr / 3000
+            bgr = 255 * bgr / 2000
             bgr = numpy.clip(bgr, 0, 255).astype("uint8")
             out.write(bgr)
+            if images:
+                im = Image.fromarray(
+                    numpy.array((bgr[:, :, 2], bgr[:, :, 1], bgr[:, :, 0])).transpose(
+                        1, 2, 0
+                    )
+                )
+                imgpath = (
+                    dst
+                    / f"videomap-{tile.z}-{tile.x}-{tile.y}-{str(img.time.data)[:10]}.png"
+                )
+                im.save(imgpath)
+
         out.release()
 
         # Convert mp4 to webm
         ffmpeg.input(str(filepath)).output(str(filepathwebm)).run(overwrite_output=True)
 
         # Delete mp4 video
-        filepath.unlink()
+        if not keep_mp4:
+            filepath.unlink()
 
         # Add tms tile feature to videos geojson object
         feat = tms.feature(tile)
